// src/types/storage.ts
export interface SettingsState {
  DarkMode: boolean;
  selectedTheme: string;
  animatedbk: boolean;
  bksliderinput: string;
  onoff: boolean;
  selectedColor: string;
  originalSelectedColor: string;
  transparencyEffects: boolean;
  justupdated?: boolean;
  animations: boolean;
  devMode?: boolean;
  originalDarkMode?: boolean;
}